conversão segundos - usp

+++ projeto-1/segundos.py
@@ -1,10 +1,10 @@
 TotalSegundos = float(input("Por favor, entre com o número de segundos que deseja converter: "))
 
-Dias = TotalSegundos // (3600*24)
-SegundoRestDias = TotalSegundos % (3600*24)
-Horas = SegundoRestDias//3600
-SegundosRestHoras = TotalSegundos % 3600
-Minutos = SegundosRestHoras // 60
-SegundosRestMinutos = SegundosRestHoras % 60
+Dias = int(TotalSegundos // (3600*24))
+SegundoRestDias = int(TotalSegundos % (3600*24))
+Horas = int(SegundoRestDias//3600)
+SegundosRestHoras = int(TotalSegundos % 3600)
+Minutos = int(SegundosRestHoras // 60)
+SegundosRestMinutos = int(SegundosRestHoras % 60)
 
 print(Dias, "dias,", Horas, "horas,", Minutos, "minutos e", SegundosRestMinutos, "segundos")
